+++ src/core/__seedwork/tests/unit/test_unit_repositories.py
@@ -1,0 +1,17 @@
+import unittest
+
+from core.__seedwork.domain.repositories import (
+    RepositoryInterface,
+)
+
+
+class TestRepositoryInterface(unittest.TestCase):
+
+    def test_throw_error_when_methods_not_implemented(self):
+        with self.assertRaises(TypeError) as assert_error:
+            # pylint: disable=abstract-class-instantiated
+            RepositoryInterface()
+        self.assertEqual(assert_error.exception.args[0],
+                        "Can't instantiate abstract class RepositoryInterface with abstract " +
+                        "methods delete, find_all, find_by_id, insert, update"
+                        )
