+++ src/django_app/asgi.py
@@ -1,0 +1,16 @@
+"""
+ASGI config for django_app project.
+
+It exposes the ASGI callable as a module-level variable named ``application``.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/4.2/howto/deployment/asgi/
+"""
+
+import os
+
+from django.core.asgi import get_asgi_application
+
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'django_app.settings')
+
+application = get_asgi_application()
